requirements.txt
python-dotenv
streamlit
langchain-community
faiss-cpu
langchain-text-splitters
langchain
langchain-core
langchain-mistralai